Add function register_APIAction_cls and variable _API_ACTION_CLS.
The class which add to _API_ACTION_CLS is going to be replaced __new__
function with add_remote_feature_to_APIAction.

When the target class init it will call __new__ method first,
than add_remote_feature_to_APIAction is going to make target class to be
competible with remote feature.